refactor: join users and auth directories
