+++ src/server/server.mts
@@ -5,7 +5,6 @@ import session from "express-session";
 import cookieParser from "cookie-parser";
 import exam from "./exam/routes.mjs";
 import media from "./media/routes.mjs";
-import auth from "./auth/routes.mjs";
 import users from "./users/routes.mjs";
 import type { Question } from "../types/globalTypes";
 
@@ -39,8 +38,8 @@ server.use(
 );
 
 server.use("/exam", exam);
-server.use("/media", media);
-server.use("/auth", auth);
+server.use("/media", media); /* 
+server.use("/auth", auth); */
 server.use("/users", users);
 
 server.listen(3001, () => {
