+++ src/server/users/routes.mts
@@ -1,8 +1,8 @@
 import Router from "express-promise-router";
-
 import { addUser, getUserByEmail } from "./users.mjs";
 import { errorMessage, errorMessageWithField } from "../messages.mjs";
-import { generateToken } from "../auth/auth.mjs";
+import { generateToken, getUserByCredentials } from "./users.mjs";
+import type { User } from "../../types/globalTypes";
 
 const router = Router();
 
@@ -29,6 +29,25 @@ router.post("/register", async (req, res) => {
     const token = generateToken(user);
 
     res.status(201).jsonp(token);
+  }
+});
+
+router.get("/login", async (req, res) => {
+  let user: User | null = null;
+  const authHeader = req.headers["authorization"];
+  if (authHeader) {
+    const credentialsCoded = authHeader.split(" ")[1];
+    const decodedCredentials = atob(credentialsCoded);
+    const email = decodedCredentials.split(":")[0];
+    const password = decodedCredentials.split(":")[1];
+
+    user = await getUserByCredentials(email, password);
+  }
+  if (user) {
+    const token = generateToken(user);
+    res.status(200).jsonp(token);
+  } else {
+    res.status(401).jsonp(errorMessage("AUTHENTICATION_FAILED"));
   }
 });
